# Reduce the time-complexity of the network initial positioning.