--- a/lib/network/modules/LayoutEngine.js
+++ b/lib/network/modules/LayoutEngine.js
@@ -161,9 +161,9 @@ class LayoutEngine {
   positionInitially(nodesArray) {
     if (this.options.hierarchical.enabled !== true) {
       this.randomSeed = this.initialRandomSeed;
+      let radius = nodesArray.length + 50;
       for (let i = 0; i < nodesArray.length; i++) {
         let node = nodesArray[i];
-        let radius = 10 * 0.1 * nodesArray.length + 10;
         let angle = 2 * Math.PI * this.seededRandom();
         if (node.x === undefined) {
           node.x = radius * Math.cos(angle);
@@ -196,33 +196,47 @@ class LayoutEngine {
       if (positionDefined < 0.5 * this.body.nodeIndices.length) {
         let MAX_LEVELS = 10;
         let level = 0;
-        let clusterThreshold = 100;
+        let clusterThreshold = 150;
+        //Performance enhancement, cluster edges need only be simpel straight lines
+        let clusterOptions = {
+          clusterEdgeProperties:{
+            smooth: {
+              enabled: false
+            }
+          }
+        };
+
         // if there are a lot of nodes, we cluster before we run the algorithm.
         if (this.body.nodeIndices.length > clusterThreshold) {
           let startLength = this.body.nodeIndices.length;
-          while (this.body.nodeIndices.length > clusterThreshold) {
+          while (this.body.nodeIndices.length > clusterThreshold && level <= MAX_LEVELS) {
             //console.time("clustering")
             level += 1;
             let before = this.body.nodeIndices.length;
             // if there are many nodes we do a hubsize cluster
             if (level % 3 === 0) {
-              this.body.modules.clustering.clusterBridges();
+              console.log("bridges");
+              this.body.modules.clustering.clusterBridges(clusterOptions);
             }
             else {
-              this.body.modules.clustering.clusterOutliers();
+              console.log("outliers");
+              this.body.modules.clustering.clusterOutliers(clusterOptions);
             }
             let after = this.body.nodeIndices.length;
-            if ((before == after && level % 3 !== 0) || level > MAX_LEVELS) {
+            if (before == after && level % 3 !== 0) {
               this._declusterAll();
               this.body.emitter.emit("_layoutFailed");
               console.info("This network could not be positioned by this version of the improved layout algorithm. Please disable improvedLayout for better performance.");
               return;
             }
             //console.timeEnd("clustering")
-            //console.log(level,after)
+            //console.log(before,level,after);
           }
           // increase the size of the edges
           this.body.modules.kamadaKawai.setOptions({springLength: Math.max(150, 2 * startLength)})
+        }
+        if (level > MAX_LEVELS){
+          console.info("The clustering didn't succeed within the amount of interations allowed, progressing with partial result.");
         }
 
         // position the system for these nodes and edges