--- a/lib/network/modules/LayoutEngine.js
+++ b/lib/network/modules/LayoutEngine.js
@@ -197,7 +197,7 @@ class LayoutEngine {
         let MAX_LEVELS = 10;
         let level = 0;
         let clusterThreshold = 150;
-        //Performance enhancement, cluster edges need only be simpel straight lines
+        //Performance enhancement, during clustering edges need only be simple straight lines. These options don't propagate outside the clustering phase.
         let clusterOptions = {
           clusterEdgeProperties:{
             smooth: {
@@ -215,11 +215,9 @@ class LayoutEngine {
             let before = this.body.nodeIndices.length;
             // if there are many nodes we do a hubsize cluster
             if (level % 3 === 0) {
-              console.log("bridges");
               this.body.modules.clustering.clusterBridges(clusterOptions);
             }
             else {
-              console.log("outliers");
               this.body.modules.clustering.clusterOutliers(clusterOptions);
             }
             let after = this.body.nodeIndices.length;