--- a/lib/network/modules/components/algorithms/FloydWarshall.js
+++ b/lib/network/modules/components/algorithms/FloydWarshall.js
@@ -4,7 +4,8 @@
 
 
 class FloydWarshall {
-  constructor(){}
+  constructor() {
+  }
 
   getDistances(body, nodesArray, edgesArray) {
     let D_matrix = {};
@@ -12,11 +13,11 @@ class FloydWarshall {
 
     // prepare matrix with large numbers
     for (let i = 0; i < nodesArray.length; i++) {
-      D_matrix[nodesArray[i]] = {};
-      D_matrix[nodesArray[i]] = {};
+      let node = nodesArray[i];
+      let cell = {};
+      D_matrix[node] = cell;
       for (let j = 0; j < nodesArray.length; j++) {
-        D_matrix[nodesArray[i]][nodesArray[j]] = (i == j ? 0 : 1e9);
-        D_matrix[nodesArray[i]][nodesArray[j]] = (i == j ? 0 : 1e9);
+        cell[nodesArray[j]] = (i == j ? 0 : 1e9);
       }
     }
 
@@ -34,10 +35,18 @@ class FloydWarshall {
 
     // Adapted FloydWarshall based on unidirectionality to greatly reduce complexity.
     for (let k = 0; k < nodeCount; k++) {
-      for (let i = 0; i < nodeCount-1; i++) {
-        for (let j = i+1; j < nodeCount; j++) {
-          D_matrix[nodesArray[i]][nodesArray[j]] = Math.min(D_matrix[nodesArray[i]][nodesArray[j]],D_matrix[nodesArray[i]][nodesArray[k]] + D_matrix[nodesArray[k]][nodesArray[j]])
-          D_matrix[nodesArray[j]][nodesArray[i]] = D_matrix[nodesArray[i]][nodesArray[j]];
+      let knode = nodesArray[k];
+      let kcolm = D_matrix[knode];
+      for (let i = 0; i < nodeCount - 1; i++) {
+        let inode = nodesArray[i];
+        let icolm = D_matrix[inode];
+        for (let j = i + 1; j < nodeCount; j++) {
+          let jnode = nodesArray[j];
+          let jcolm = D_matrix[jnode];
+
+          let val = Math.min(icolm[jnode], icolm[knode] + kcolm[jnode]);
+          icolm[jnode] = val;
+          jcolm[inode] = val;
         }
       }
     }