Very substantial performance gain for large graphs. (>500 nodes) 
Price of this patch is more complex code in an already complex part, but reducing large O complexity is important, for large graphs this reduces the rendering times substantially.