@mojoaxel 2017-02-21
I made a few simple performance tests with this options:
```js
{
  physics: {
    stabilization: {
      enabled: true,
      iterations: 10000,
      updateInterval: 100,
      onlyDynamicEdges: false,
      fit: true
    }
  },
  layout: {
    randomSeed: 123
  },
  edges:{
    smooth:{
      enabled: false,
      type:'continuous'
    }
  },
}
```
I got the following results:

**Here less/smaller is better!**

![image](https://cloud.githubusercontent.com/assets/600565/23182454/dbac4588-f878-11e6-9b3a-981dcaf9196e.png)

![image](https://cloud.githubusercontent.com/assets/600565/23182525/1dd32d5a-f879-11e6-8e07-8135ec16537b.png)

![image](https://cloud.githubusercontent.com/assets/600565/23182497/fe07f56e-f878-11e6-8596-e8f5720ef51d.png)

I have to say I only ran the test one time and strong deviations (e.g. at 450 nodes) is propably due to random circumstances.

It looks like this changes work really good for small networks (<300). 

But you can also see that there are also a significant increase in render time for 600<nodes<850 and 300 and 400.  I have no Idea why, but whe should maybe fix this before merging this.

If needed I could write some performance tests that run automatically. I could run these e.g. overnight on my 12 core workstation. But I'm not sure If I find the time to get to it this week.