@ludost 2017-02-22
Hi, 
Thanks for your feedback and review. I'll fix the console messages, this change was in my working copy for a while and I forgot to remove them while merging with the newest development branch.

Re performance: I think this is mostly due to the higher memory usage in my patch. If this triggers a substantial GC round, this might explain the non-linear behavior in growing numbers of nodes. I'll redo this patch watching that behavior more closely, checking if that explains your numbers.

Good to see that there are more performance patches. I suggest merging #2729 first and letting me update this to match that new situation. (and hopefully fix this for the whole range)

Kind regards,
Ludo.