@mojoaxel 2017-02-22
> I suggest merging #2729 first and letting me update this to match that new situation. (and hopefully fix this for the whole range)

Perfect. Thanks you very much. That makes it much easier!