@ludost 2017-02-27
Hi Alexander,

Can you provide some more details about those performance tests? You've provided the options, and I assume you used the performance example to generate the network? What is your end condition and how do you measure it? 

I like to run this and similar tests (e.g. a dense, highly-connected graph) myself, and wouldn't like to do double work:)

Thanks!
Greetings,
Ludo.